requirements.txt
aws-cdk-lib==2.0.0-rc.33
constructs>=10.0.0,<11.0.0
